quick settings pulldown, left or right side option
Quick pulldown sometimes flips when it should not

The new code simply checks to see that the Event was a pull down event
and that the event started in the top status bar.
If this is the case, then the quick panel is swapped to (without flip)

Special Thanks to my daughter Jenny fo painstakingly reproducing this
bug.

Fix the flipping animation sometimes being played when pulling the quick
settings with two fingers

Finally fix incorrect flipping to quick settings